Gui Expansion Version + Bug Fixes
I took the liberty of giving the application a GUI, and provided a way for the server to deny file uploads via a pop-up.

I discovered also that sending large files didn't work since I relied upon InputStream.available(), which only returns the # of bytes you can read before it blocks, not the actual size of the file.
This was easy to fix, although the actual transfer of the bytes is slow.
Allegedly there would be issues with making a byte array for a large file, so I chose to send one byte at a time and read one at a time. This makes the overall process much slower, but it is more scalable I guess.

To-do:
Implement threads, and the ability for the server to start receiving connections or to refuse them, and it should display their IP when they start the server.
Part of this change would require the ServerSocket object to not be discarded until the server is closed, or until the application itself is closed (the window)